Replaced nNumberEpochs by nMaxEpochs, changed int to long for rows of files, made a routine to initialize all trainings like linear regression to save time, nMaxEpochs is now the number of epochs between splittings, simplified do splitcontrol, corrected pALN->SetWeightMax / Min to have 3.0 in place of 2.0. It runs but overtraining error is far from zero.

+++ libaln/include/alnpriv.h
@@ -163,7 +163,7 @@ void ALNAPI DebugValidateALNDataInfo(const ALN* pALN,
 
 
 // calculate start and end points of data set given varinfo deltas
-void ALNAPI CalcDataEndPoints(int& nStart, int& nEnd, 
+void ALNAPI CalcDataEndPoints(long& nStart, long& nEnd, 
                               const ALN* pALN,
                               const ALNDATAINFO* pDataInfo);
